test/inspector/debugger/set-script-source-exception-expected.txt

Check that setScriptSource completes correctly when an exception is thrown.

Running test: testIncorrectScriptId
{
    error : {
        code : -32000
        message : No script with given id found
    }
    id : <messageId>
}

Running test: testSourceWithSyntaxError
{
    error : {
        code : -32000
        message : Uncaught [object Object]
    }
    id : <messageId>
}
